adding FHSA calc in and reorg
