+++ apps/frontend/src/_components/Calculator-Component/FHSA/FhsaResultComponent.tsx
